Fixed duplication bug
For whatever reason, this plugin was creating a new span every time keydown was triggered which is not good for performance. So instead I added an id to the span you create and add to body and have a check to see if it exists. If it does I simply use that existing span instead of creating a new one every time.